requirements.txt
apiclient==1.0.4
appnope==0.1.3
asttokens==2.4.1
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
comm==0.2.0
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
gspread==5.12.0
idna==3.4
ipykernel==6.26.0
ipython==8.17.2
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.5.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.26.2
oauthlib==3.2.2
packaging==23.2
pandas==2.1.3
parso==0.8.3
pexpect==4.8.0
platformdirs==4.0.0
prompt-toolkit==3.0.41
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.16.1
python-dateutil==2.8.2
pytz==2023.3.post1
pyzmq==25.1.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
stack-data==0.6.3
tornado==6.3.3
traitlets==5.13.0
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.10
